task 2 part 18

+++ alembic/versions/initial_migration.py
@@ -9,8 +9,8 @@ depends_on = None
 def upgrade():
     op.create_table(
         'user_values',
-        sa.Column('id', sa.Integer(), nullable=False),
-        sa.Column('user_id', sa.Integer(), nullable=False),
+        sa.Column('id', sa.BigInteger(), nullable=False),
+        sa.Column('user_id', sa.BigInteger(), nullable=False),
         sa.Column('value', sa.String(length=255), nullable=False),
         sa.Column('created_at', sa.DateTime(), nullable=True),
         sa.PrimaryKeyConstraint('id')
