+++ models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, BigInteger, String, DateTime  # Добавлен Integer
+from sqlalchemy import Column, Integer, BigInteger, String, DateTime
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import mapped_column
 from datetime import datetime
@@ -8,7 +8,7 @@ Base = declarative_base()
 class UserValue(Base):
     __tablename__ = "user_values"
     
-    id = mapped_column(Integer, primary_key=True)
+    id = mapped_column(BigInteger, primary_key=True)  
     user_id = mapped_column(BigInteger, nullable=False)
     value = mapped_column(String(255), nullable=False)
     created_at = mapped_column(DateTime, default=datetime.utcnow)
